Remove old orion 'client' code
