--- experiments/comet/docker-compose.yml
@@ -1,38 +1,0 @@
-version: '3'
-
-services:
-
-  orion:
-    image: fiware/orion:1.7.0
-    ports:
-      - "1026:1026"
-    command: -logLevel DEBUG -dbhost mongo
-    depends_on:
-      - mongo
-    healthcheck:
-      test: ["CMD", "curl", "-f", "http://0.0.0.0:1026/version"]
-      interval: 1m
-      timeout: 10s
-      retries: 3
-
-  mongo:
-    image: mongo:3.2
-    ports:
-      - "27017:27017"
-
-  comet:
-    image: telefonicaiot/fiware-sth-comet:2.2.0
-    ports:
-      - "8666:8666"
-    environment:
-      - STH_HOST=0.0.0.0
-      - DB_URI=mongo:27017
-      - LOGOPS_LEVEL=DEBUG
-    depends_on:
-      - orion
-      - mongo
-    healthcheck:
-      test: ["CMD", "curl", "-f", "http://0.0.0.0:8666/version"]
-      interval: 2m
-      timeout: 10s
-      retries: 3
