--- experiments/grafana/docker-compose.yml
@@ -1,63 +1,0 @@
-version: '3'
-
-services:
-
-  orion:
-    image: fiware/orion:1.7.0
-    ports:
-      - "1026:1026"
-    command: -logLevel DEBUG -noCache -dbhost mongo
-    depends_on:
-      - mongo
-    healthcheck:
-      test: ["CMD", "curl", "-f", "http://0.0.0.0:1026/version"]
-      interval: 1m
-      timeout: 10s
-      retries: 3
-
-  mongo:
-    image: mongo:3.2
-    ports:
-      - "27017:27017"
-    volumes:
-      - mongodata:/data/db
-
-  quantumleap:
-    image: smartsdk/quantumleap
-    ports:
-      - "8668:8668"
-    depends_on:
-      - mongo
-      - orion
-      - crate
-    environment:
-      - CRATE_HOST=crate
-
-  crate:
-    image: crate:1.0.5
-    ports:
-      # Admin UI
-      - "4200:4200"
-      # Transport protocol
-      - "4300:4300"
-    command: -Ccluster.name=democluster -Chttp.cors.enabled=true -Chttp.cors.allow-origin="*"
-    volumes:
-      - cratedata:/data
-
-  grafana:
-    image: grafana/grafana
-    ports:
-      - "3000:3000"
-    environment:
-      - GF_INSTALL_PLUGINS=crate-datasource,grafana-clock-panel,grafana-worldmap-panel
-    depends_on:
-      - crate
-
-volumes:
-  mongodata:
-  cratedata:
-
-networks:
-    default:
-        driver_opts:
-            com.docker.network.driver.mtu: ${DOCKER_MTU:-1400}
