+++ reporter/tests/test_1T1E1A.py
@@ -1,4 +1,3 @@
-from client.client import HEADERS
 from conftest import QL_URL
 from reporter.tests.utils import insert_test_data
 from translators.fixtures import crate_translator as translator
@@ -37,7 +36,7 @@ def test_1T1E1A_defaults(reporter_dataset):
     query_params = {
         'type': entity_type,
     }
-    r = requests.get(query_url(), params=query_params, headers=HEADERS)
+    r = requests.get(query_url(), params=query_params)
     assert r.status_code == 200, r.text
 
     # Assert
@@ -69,7 +68,7 @@ def test_1T1E1A_aggrMethod(reporter_dataset, aggr_method, aggr_value):
         'type': entity_type,
         'aggrMethod': aggr_method,
     }
-    r = requests.get(query_url(), params=query_params, headers=HEADERS)
+    r = requests.get(query_url(), params=query_params)
     assert r.status_code == 200, r.text
 
     # Assert
@@ -91,7 +90,7 @@ def test_1T1E1A_fromDate_toDate(reporter_dataset):
         'fromDate': "1970-01-06T00:00:00",
         'toDate': "1970-01-17T00:00:00",
     }
-    r = requests.get(query_url(), params=query_params, headers=HEADERS)
+    r = requests.get(query_url(), params=query_params)
     assert r.status_code == 200, r.text
 
     # Expect only last N
@@ -121,7 +120,7 @@ def test_1T1E1A_lastN(reporter_dataset):
         'type': entity_type,
         'lastN': 10
     }
-    r = requests.get(query_url(), params=query_params, headers=HEADERS)
+    r = requests.get(query_url(), params=query_params)
     assert r.status_code == 200, r.text
 
     # Expect only last N
@@ -151,7 +150,7 @@ def test_1T1E1A_limit(reporter_dataset):
         'type': entity_type,
         'limit': 5
     }
-    r = requests.get(query_url(), params=query_params, headers=HEADERS)
+    r = requests.get(query_url(), params=query_params)
     assert r.status_code == 200, r.text
 
     # Expect only last N
@@ -181,7 +180,7 @@ def test_1T1E1A_offset(reporter_dataset):
         'type': entity_type,
         'offset': 3
     }
-    r = requests.get(query_url(), params=query_params, headers=HEADERS)
+    r = requests.get(query_url(), params=query_params)
     assert r.status_code == 200, r.text
 
     # Expect only last N
@@ -213,7 +212,7 @@ def test_1T1E1A_combined(reporter_dataset):
         'toDate': "1970-01-20T00:00:00",
         'limit': 28,
     }
-    r = requests.get(query_url(), params=query_params, headers=HEADERS)
+    r = requests.get(query_url(), params=query_params)
     assert r.status_code == 200, r.text
 
     # Expect only last N
@@ -242,9 +241,7 @@ def test_1T1E1A_values_defaults(reporter_dataset):
     query_params = {
         'type': entity_type,
     }
-    r = requests.get(query_url(values=True),
-                     params=query_params,
-                     headers=HEADERS)
+    r = requests.get(query_url(values=True), params=query_params)
     assert r.status_code == 200, r.text
 
     # Assert
@@ -265,7 +262,7 @@ def test_not_found():
     query_params = {
         'type': entity_type,
     }
-    r = requests.get(query_url(), params=query_params, headers=HEADERS)
+    r = requests.get(query_url(), params=query_params)
     assert r.status_code == 404, r.text
     assert r.json() == {
         "error": "Not Found",
@@ -277,7 +274,7 @@ def test_tmp_no_type():
     """
     For now specifying entity type is mandatory
     """
-    r = requests.get(query_url(), params={}, headers=HEADERS)
+    r = requests.get(query_url(), params={})
     assert r.status_code == 400, r.text
     assert r.json() == {
         "error": "Not Implemented",
