+++ reporter/tests/test_subscribe.py
@@ -1,5 +1,4 @@
-from client.fixtures import clean_mongo
-from conftest import QL_URL, ORION_URL
+from conftest import QL_URL, ORION_URL, clean_mongo, clean_crate
 import requests
 
 subscribe_url = "{}/subscribe".format(QL_URL)
