+++ run_tests.sh
@@ -2,17 +2,9 @@
 
 docker build -t quantumleap .
 
-cd client/tests
-sh run_tests.sh
-tot=$?
-cd -
-
 cd translators/tests
 sh run_tests.sh
-loc=$?
-if [ "$tot" -eq 0 ]; then
-   tot=$loc
-fi
+tot=$?
 cd -
 
 cd reporter/tests
